refactor: split ui facade and query output bands
